% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Title}
\usage{
estimate_cate(
  trial_tbl,
  estimation_method = c("slearner", "xlearner", "causalforest"),
  aggregation_method = c("studyindicator", "ensembleforest"),
  site_col,
  treat_col,
  outcome_col,
  covariate_col = NULL,
  drop_col = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
