% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_aug_tau.R
\name{generate_aug_tau}
\alias{generate_aug_tau}
\alias{generate_aug_tau.causalforest_args}
\alias{generate_aug_tau.slearner_args}
\title{Generate Augmented Dataset Based on Selected Estimation Method}
\usage{
generate_aug_tau(named_args, study_val, ...)

\method{generate_aug_tau}{causalforest_args}(named_args, study_val, ...)

\method{generate_aug_tau}{slearner_args}(named_args, study_val, ...)
}
\arguments{
\item{named_args}{list. A list of pre-specified argument values from
\code{multicate::estimate_cate()}.}

\item{study_val}{string, numeric, or factor. Unique identifier for primary study of interest}

\item{...}{Additional arguments to be passed to \code{estimation_method}.}
}
\value{
A tbl with additional column \code{tau_hat} predicted for each observation based on a model
using the selected \code{estimation_method} trained on each study in turn. As a result, every
observation will have a separate row with a tau_hat prediction based on models trained on each
study (e.g., if the original data includes k studies, each with n observations, then the returned
tbl will have nk observations)
}
\description{
Generate Augmented Dataset Based on Selected Estimation Method
}
