% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_tau.R
\name{generate_tau.studyindicator_args}
\alias{generate_tau.studyindicator_args}
\title{Title}
\usage{
\method{generate_tau}{studyindicator_args}(named_args, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
