% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cate.R
\name{plot.cate}
\alias{plot.cate}
\title{Title}
\usage{
\method{plot}{cate}(
  cate_obj,
  which_plot = 1:3,
  captions = c("Histogram of CATEs", "Boxplot of CATEs vs Study Membership",
    "Confidence Intervals for CATEs"),
  ask = TRUE
)
}
\arguments{
\item{captions}{}
}
\description{
Title
}
