% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cate.R
\name{plot_blp}
\alias{plot_blp}
\title{Plot Best Linear Projection}
\usage{
plot_blp(cate_obj)
}
\arguments{
\item{cate_obj}{list. An object resulting from \link{estimate_cate}.}
}
\description{
Plot Best Linear Projection
}
\details{
Additional details...
}
