% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cate.R
\name{plot_vteffect}
\alias{plot_vteffect}
\title{Plot Variable Treatment Effect}
\usage{
plot_vteffect(object, covariate_name)
}
\arguments{
\item{object}{list. An object resulting from \link{estimate_cate}.}

\item{covariate_name}{string. Name of a covariate included in dataset used to estimate tau_hat.}
}
\description{
Plot Variable Treatment Effect
}
\details{
This function plots the value of the selected covariate for each observation in the dataset
against the value of tau_hat for the variable. This is what the findings mean...
}
