% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cate.R
\name{predict.cate}
\alias{predict.cate}
\title{Predict Heterogeneous Treatment Effects Based on Estimated CATEs}
\usage{
\method{predict}{cate}(cate_obj, newdata_tbl, alpha = 0.05)
}
\arguments{
\item{cate_obj}{list. An object resulting from \link{estimate_cate}.}

\item{newdata_tbl}{tbl. A tbl containing columns for treatment, outcome, study ID, and any
additional covariates used to generate \code{cate_obj}. All study data must be included in single
tbl. Note that only two treatments can be considered and treatment must be coded as 0/1
(numeric).}

\item{alpha}{numeric. Significance level for confidence intervals. Defaults to 0.05 for 95\% CIs.}
}
\value{
A list
}
\description{
Predict Heterogeneous Treatment Effects Based on Estimated CATEs
}
