% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_build_augmented_data.R
\name{predict_tau.causalforest_args}
\alias{predict_tau.causalforest_args}
\title{Title}
\usage{
\method{predict_tau}{causalforest_args}(named_args, site_val, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
