% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_cate.R
\name{summary.cate}
\alias{summary.cate}
\alias{print.summary.cate}
\title{Summarizing CATE Model Fits}
\usage{
\method{summary}{cate}(object, tauhat_column = "tau_hat", ...)

\method{print}{summary.cate}(x, ...)
}
\arguments{
\item{object}{list. An object of class "cate", usually a result of a call to \link{estimate_cate}.}

\item{...}{Additional arguments passed to or from other methods.}

\item{x}{list. An object of class \code{summary.cate}.}

\item{tauhat_col}{string. When \code{aggregation_method} is "ensembleforest", name of the column in
\code{object$model} that has the estimated CATEs. Default is "tau_hat".}
}
\value{
A list of summary statistics and tables based on the information given in \code{object},
including
\itemize{
\item \verb{estimation method} : The estimation method used to estimate the CATEs.
\item \verb{aggregation method} : The aggregation method used to estimate the CATEs.
\item \code{vi_table} : The variable importance table.
\item \code{ate} : The overall average treatment effect across all patients and all studies. When the
\code{estimation_method} is "causalforest", this is estimated using the
\link[grf:average_treatment_effect]{grf::average_treatment_effect} function. When the
\code{estimation_method} is "slearner", this is estimated by averaging the treatment effect across
patients, accounting for counterfactual assignments. When the \code{aggregation_method} is
"ensembleforest", a simple arithmetic mean of estimated CATEs is provided and the standard
error is \code{NA}.
\item \code{studycate} : A table of the minimum, median, and maximum CATE values for patients from each
study included in the model.
}
}
\description{
Summarizing CATE Model Fits
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(summary.cate)}: Print the summary of a CATE model

}}
\examples{
# inst/examples/example-summary_cate.R

cate_object <- estimate_cate(
  trial_tbl = dummy_tbl,
  estimation_method = "causalforest",
  aggregation_method = "studyindicator",
  study_col = "studyid",
  treatment_col = "tx",
  outcome_col = "response"
)

summary(cate_object)
}
