% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{theme_MH}
\alias{theme_MH}
\title{Custom ggplot2 Theme}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 11.
}
\usage{
theme_MH
}
\description{
A reusable custom ggplot2 theme for consistent styling.
}
\keyword{datasets}
