% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Estimate Heterogeneous Treatment Effects Across Multiple Randomized Control Trials}
\usage{
estimate_cate(
  trial_tbl,
  estimation_method = c("slearner", "xlearner", "causalforest"),
  aggregation_method = c("studyindicator", "ensembleforest"),
  study_col,
  treatment_col,
  outcome_col,
  covariate_col = NULL,
  drop_col = NULL,
  incl_cfobject = TRUE,
  ...
)
}
\arguments{
\item{trial_tbl}{tbl. A tbl containing columns for treatment, outcome, study ID, and any
additional covariates of interest. All study data must be included in single tbl. Note that
only two treatments can be considered and treatment must be coded as 0/1 (numeric).}

\item{estimation_method}{string. Single-study methods for estimating CATE (tau) for each
observation. Available methods are "slearner" (using Bayesian Additive Regression Trees),
"xlearner", and "causalforest".}

\item{aggregation_method}{string. Method for aggregating results across studies Available methods
are:
\itemize{
\item "studyindicator": Pool all data together but keep study ID as an indicator variable and
includes them as covariates in the single-study method selected in \code{estimation_method}.
\item "ensembleforest": Fit the \code{estimation_method} within each study, apply each model to all
individuals across all studies, then fit ensemble random forest to augmented data.
}}

\item{study_col}{string. Name of the column in \code{trial_tbl} with study ID.}

\item{treatment_col}{string. Name of the column in \code{trial_tbl} with treatment assignment.}

\item{outcome_col}{string. Name of the column in \code{trial_tbl} with outcome values.}

\item{covariate_col}{character vector. Name(s) of columns in \code{trial_tbl} to be included as
model covariates. Defaults to NULL.}

\item{drop_col}{character vector. Name(s) of columns in \code{trial_tbl} to be excluded from
model covariates. Defaults to NULL.}

\item{incl_cfobject}{logical. When TRUE, if \code{estimation_method} is "causalforest" the causal
forest object is saved and returned. Defaults to TRUE.}

\item{...}{Arguments to be passed to \code{estimation_method} and/or \code{aggregation_method}. Note the
following exceptions:
\itemize{
\item \code{estimation_method} = "slearner"
\itemize{
\item \link[dbarts:bart]{dbarts::bart()} argument \code{keeptrees} set to TRUE
\item \link[dbarts:bart]{dbarts::bart()} argument \code{verbose} set to FALSE when \code{aggregation_method} is
"ensembleforest"
}
\item \code{aggregation_method} = "ensembleforest"
\itemize{
\item \link[ranger:ranger]{ranger::ranger()} argument \code{importance} set to "impurity"
\item \link[ranger:ranger]{ranger::ranger()} argument \code{keep.inbag} set to TRUE
}
}}
}
\value{
An object of class "cate" is a list containing the following elements:
\itemize{
\item \code{estimation_method}: the type of estimation method used
\item \code{aggregation_method}: the type of aggregation method used
\item \code{model}: \code{trial_tbl} with additional columns
\itemize{
\item \code{tau_hat}: conditional average treatment effect for each observation
\item \code{variance_estimates}: variance estimate for each tau_hat (not available when
\code{estimation_method} set to "xlearner" or \code{aggregation_method} set to "ensembleforest)
}
\item \code{var_importance}: TBD
\item \code{study_col}: name of study ID column
\item \code{treatment_col}: name of treatment column
\item \code{outcome_col}: name of outcome column
\item \code{covariate_col}: name(s) of covariate columns
\item \code{causalforest}: Trained causal forest object from \link[grf:causal_forest]{causal_forest} (If
\code{estimation_method} is set to "causalforest" and \code{incl_cfobject} is TRUE)
}
}
\description{
\code{estimate_cate} is used to estimate conditional average treatment effect (CATE) from multiple
studies.
}
