% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_cate.R
\name{estimate_cate}
\alias{estimate_cate}
\title{Estimate Heterogeneous Treatment Effects Across Multiple Randomized Control Trials}
\usage{
estimate_cate(
  trial_tbl,
  estimation_method = c("slearner", "xlearner", "causalforest"),
  aggregation_method = c("studyindicator", "ensembleforest"),
  site_col,
  treatment_col,
  outcome_col,
  covariate_col = NULL,
  drop_col = NULL,
  incl_cfobject = FALSE,
  ...
)
}
\arguments{
\item{trial_tbl}{tbl. A tbl containing columns for treatment, outcome, site ID, and any
additional covariates of interest. All site data must be included in single tbl.}

\item{estimation_method}{string. Single-study mehtods for estimating CATE (tau) for each
observation. Available methods are "slearner" (using Bayesian Additive Regression Trees),
"xlearner", and "causalforest".}

\item{aggregation_method}{string. Method for aggregating results across sites. Available methods
are:
\itemize{
\item "studyindicator": Pool all data together but keep study ID as an indicator variable and
includes them as covariates in the single-study method selected in \code{estimation_method}.
\item "ensembleforest": Fit the \code{estimation_method} within each study, apply each model to all
individuals across all sites, then fit ensemble random forest to augmented data.
}}

\item{site_col}{string. Name of the column in \code{trial_tbl} with site ID.}

\item{treatment_col}{string. Name of the column in \code{trial_tbl} with treatment assignment.}

\item{outcome_col}{string. Name of the column in \code{trial_tbl} with outcome values.}

\item{covariate_col}{character vector. Name(s) of columns in \code{trial_tbl} to be included as
model covariates. Defaults to NULL.}

\item{drop_col}{character vector. Name(s) of columns in \code{trial_tbl} to be excluded from
model covariates. Defaults to NULL.}

\item{incl_cfobject}{logical. When TRUE, if \code{estimation_method} is "causalforest" the causal
forest object is saved and returned.}

\item{...}{Arguments to be passed to \code{estimation_method} and/or \code{aggregation_method}. Note the
following exceptions:
\itemize{
\item \code{estimation_method} = "slearner"
\itemize{
\item \code{dbarts::bart()} argument \code{keeptrees} set to TRUE
\item \code{dbarts::bart()} argument \code{verbose} set to FALSE when \code{aggregtion_method} is
"ensembleforest"
}
\item \code{aggregation_method} = "ensembleforest"
\itemize{
\item \code{ranger::ranger()} argument \code{importance} set to "impurity"
\item \code{ranger::ranger()} argument \code{keep.inbag} set to TRUE
}
}}
}
\value{
An object of class "cate" is a list containing the following components:
\itemize{
\item \code{estimation_method}: the type of estimation method used
\item \code{aggregation_method}: the type of aggregation method used
\item \code{model}: \code{trial_tbl} with additional columns
\itemize{
\item \code{tau_hat}: conditional average treatment effect for each observation
\item \code{variance_estimates}: variance estimate for each tau_hat (not available when
\code{estimation_method} set to "xlearner" or \code{aggregation_method} set to "ensembleforest)
}
\item \code{var_importance}: TBD
\item \code{site_col}: name of site ID column
\item \code{treatment_col}: name of treatment column
\item \code{outcome_col}: name of outcome column
\item \code{covariate_col}: name(s) of covariate columns
}
}
\description{
\code{estimate_cate} is used to estimate conditional average treatment effect (CATE) from multiple
studies.
}
