% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_tau.R
\name{generate_tau}
\alias{generate_tau}
\title{Title}
\usage{
generate_tau(
  feature_tbl,
  site_col,
  treatment_vec,
  outcome_vec,
  estimation_method,
  aggregation_method,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
Title
}
