% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cate.R
\name{plot.cate}
\alias{plot.cate}
\title{Plot Diagnostics for a CATE Object}
\usage{
\method{plot}{cate}(x, which_plot = 1:5, ask = TRUE, ...)
}
\arguments{
\item{x}{list. An object resulting from \link{estimate_cate}.}

\item{which_plot}{numeric vector. A vector indicating which plots should be generated.}

\item{ask}{logical. When TRUE, the user is asked before each plot, see
\link[graphics:par]{par(ask = .)}.}

\item{...}{Additional arguments passed to or from other methods.}
}
\description{
Five plots (selectable by \code{which_plot}) currently available: a histogram of estimated conditional
average treatment effects (CATEs), a boxplot of CATEs stratified by study membership, a plot of
95\% confidence intervals for all CATEs sorted in order to CATE estimate value, a best linear
projection (only available when \code{estimation_method} is set to "causalforest"), and an
interpretation tree.
}
\examples{
# inst/examples/example-plot_cate.R

cate_object <- estimate_cate(
  trial_tbl = dummy_tbl,
  estimation_method = "causalforest",
  aggregation_method = "studyindicator",
  study_col = "studyid",
  treatment_col = "tx",
  outcome_col = "response"
)

plot(cate_object)

## Use `which_plot` to select specific plots to produce
## Use `ask = FALSE` to automatically generate all plots
plot(cate_object,
     which_plot = c(1, 3, 5),
     ask = FALSE)
}
