% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_cate.R
\name{plot.cate}
\alias{plot.cate}
\title{Plot Diagnostics for a CATE Object}
\usage{
\method{plot}{cate}(cate_obj, which_plot = 1:3, ask = TRUE)
}
\arguments{
\item{cate_obj}{\code{cate} list. An object resulting from \link{estimate_cate}.}

\item{which_plot}{numeric vector. A vector indicating which plots should be generated.}

\item{ask}{logical. When TRUE, the user is asked before each plot, see
\link[graphics:par]{par(ask = .)}.}
}
\description{
Four plots (selectable by \code{which_plot}) currently available: a histogram of estimated conditional
average treatment effects (CATEs), a boxplot of CATEs stratified by study membership, a plot of
95\% confidence intervals for all CATEs sorted in order to CATE estimate value, and an
interpretation tree (only available when \code{estimation_method} is set to "causalforest")
}
