% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cate.R
\name{predict.cate}
\alias{predict.cate}
\title{Predict Heterogeneous Treatment Effects Based on Estimated CATEs}
\usage{
\method{predict}{cate}(object, newdata_tbl, alpha = 0.05, ...)
}
\arguments{
\item{object}{list. An object resulting from \link{estimate_cate}.}

\item{newdata_tbl}{tbl. A tbl containing columns for treatment, outcome, study ID, and any
additional covariates used to generate \code{cate_obj}. All study data must be included in single
tbl. Note that only two treatments can be considered and treatment must be coded as 0/1
(numeric).}

\item{alpha}{numeric. Significance level for confidence intervals. Defaults to 0.05 for 95\% CIs.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A tbl including all columns and data from \code{newdata_tbl}, with additional columns
including a mean predicted CATE and associated 100(1 - \eqn{\alpha})\% prediction interval.
}
\description{
\code{predict.cate} is used to produce a prediction of the conditional average treatment effect (CATE)
based on models fit to data from multiple studies/sites.
}
\details{
The predict function uses CATE models fit to multiple studies (objects resulting from
\link{estimate_cate}) to form prediction intervals for CATEs in a new, target setting (e.g., a
group of patients in an electronic health record system). This function leverages two-stage meta-
analysis with prediction intervals for all covariate profiles (\eqn{\mathbf{X}^*}) of interest in
the target setting.

The first stage of the two-stage meta-analysis is completed by fitting models to each study individually
to estimate the CATE per study using the \link{estimate_cate} function with
the "studyspecific" aggregation method. From this first stage, we can obtain estimates of the CATE for a
given covariate profile in the target setting in each study i: \eqn{\hat{\tau}_i(\mathbf{X}^*)}, as well as variance
estimates of those CATEs.

The second stage of the two-stage meta-analysis is replicated for every covariate profile
\eqn{\mathbf{X}^*} of interest. In this stage, we fit a random effects meta-analysis under the
distributional assumptions:

\deqn{\hat{\tau}_i(\mathbf{X}^*) \sim
N(\tau_i(\mathbf{X}^*),
{s_i}^2(\mathbf{X}^*))}

where \eqn{\hat{\tau}_i(\mathbf{X}^*)}
is the estimated CATE for \eqn{\mathbf{X}^*} in study i,
\eqn{\tau_i(\mathbf{X}^*)} is the true CATE for
\eqn{\mathbf{X}^*} in study i, and
\eqn{{s_i}^2(\mathbf{X}^*)} is the within-study variability estimate of the CATE for
\eqn{\mathbf{X}^*}; and

\deqn{\tau_i(\mathbf{X}^*) \sim N(\tau(\mathbf{X}^*), \theta^2(\mathbf{X}^*))}

where \eqn{\tau(\mathbf{X}^*)}
is the true average CATE for \eqn{\mathbf{X}^*} across all studies, and
\eqn{\theta^2(\mathbf{X}^*)} is the between-study variance.

From this two-stage meta-analysis, the prediction interval for the CATE for covariate profile
\eqn{\mathbf{X}^*} in the target setting represents a range of potential values that the CATE
may be in the new setting. This prediction interval is of the form:

\deqn{\hat{\tau}(\mathbf{X}^*) \pm t_{K-2} \sqrt{SE(\hat{\tau}(\mathbf{X}^*)^2 +
\hat{\theta}^2(\mathbf{X}^*)}}

where \eqn{\hat{\tau}(\mathbf{X}^*)} is the inverse variance-weighted average of the
\eqn{\hat{\tau}_i(\mathbf{X}^*)} across studies, \eqn{t_{K-2}} is the t-statistic with K-2 degrees
of freedom (where K is the total number of studies), \eqn{SE(\hat{\tau}(\mathbf{X}^*)^2} is the
estimated variance of \eqn{\hat{\tau}(\mathbf{X}^*)}, and \eqn{\hat{\theta}^2(\mathbf{X}^*)} is
the estimated between-study variance, calculated using REML. For more details, see Brantner et
al., (Under review) and Riley et al. (2021).

The predict function can only be run on \code{cate} objects using the "studyspecific" aggregation
method. Any estimation method (i.e., causal forest or S-learner with BART) can be used.

The \code{newdata_tbl} can consist of any set of profiles for which the researcher would like to
create prediction intervals. The assumption with this approach is that the profiles in the target
setting can be found within the covariate distribution of at least one of the studies used to
estimate the CATE (i.e., positivity). CATE prediction intervals will be constructed separately
for each covariate profile, represented by a row in the newdata_tbl.
}
\examples{
# inst/examples/example-predict_cate.R

new_dummy_tbl <- tibble::tribble(
  ~tx, ~var1,   ~var2,   ~var3,    ~var4,   ~var5,   ~response,
  0,  -0.728,   1.66,   -0.213,   -0.0621,  0.252,   1.26,
  0,  -0.464,   0.230,   0.579,   -1.17,    0.0513, -0.235,
  0,   1.52,   -0.792,   0.951,    0.869,  -0.312,   1.42,
  0,   0.410,  -0.945,  -1.49,    -1.87,   -1.35,   -3.10,
  1,  -0.223,   2.20,    0.0242,   0.419,   0.333,   3.56,
  1,  -0.187,  -0.439,  -0.557,   -0.0679,  0.497,   0.0518,
  1,   1.41,   -0.830,  -0.306,    1.12,    0.648,   2.61,
  1,  -0.437,   1.22,    1.18,    -1.63,   -1.09,    0.833,
)

cate_object <- estimate_cate(
  trial_tbl = dummy_tbl,
  estimation_method = "causalforest",
  aggregation_method = "studyspecific",
  study_col = "studyid",
  treatment_col = "tx",
  outcome_col = "response"
)

predict(cate_object, new_dummy_tbl)
}
\references{
Brantner CL, Nguyen TQ, Parikh H, Zhao C, Hong H, Stuart EA. Precision mental health: Predicting
heterogeneous treatment effects for depression through data integration. \emph{Under review.}

Riley, RD, Debray TP, Morris TP, Jackson D. The Two-Stage Approach to IPD Meta-Analysis. In
Individual Participant Data Meta-Analysis (pp 87-125). (2021)
}
