% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cate.R
\name{predict.cate}
\alias{predict.cate}
\title{Predict Heterogeneous Treatment Effects Based on Estimated CATEs}
\usage{
\method{predict}{cate}(object, newdata_tbl, alpha = 0.05, ...)
}
\arguments{
\item{object}{list. An object resulting from \link{estimate_cate}.}

\item{newdata_tbl}{tbl. A tbl containing columns for treatment, outcome, study ID, and any
additional covariates used to generate \code{cate_obj}. All study data must be included in single
tbl. Note that only two treatments can be considered and treatment must be coded as 0/1
(numeric).}

\item{alpha}{numeric. Significance level for confidence intervals. Defaults to 0.05 for 95\% CIs.}

\item{...}{Additional arguments passed to or from other methods.}
}
\value{
A tbl including all columns and data from \code{newdata_tbl}, with additional columns
including a mean predicted CATE and associated 100(1 - \eqn{\alpha})\% prediction interval.
}
\description{
\code{predict_cate} is used to produce a prediction of the conditional average treatment effect (CATE)
based on a model built on multiple studies.
}
\details{
The predict function uses estimated CATEs from multiple studies (objects resulting from
\link{estimate_cate}) to form prediction intervals for CATEs in a new, target setting (e.g., a
group of patients in an electronic health record system). This function leverages prediction
interval techniques from meta-analysis by estimating the within-study and between-study variance
of the CATEs for all covariate profiles (X) of interest in the target setting. The prediction
interval for the CATE for covariate profile X in the target setting is of the form:
\deqn{\hat{\tau}(\mathbf{X}^*) \pm \sqrt{var_{within} + var_{between}}}
Where \eqn{\hat{\tau}(\mathbf{X}^*)} is the inverse variance-weighted average of the
\eqn{\hat{\tau}(\mathbf{X}^*)} across studies, \eqn{t_{K-2}} is the t-statistic with K-2 degrees
of freedom, var_within is the average of the variances of each \eqn{\hat{\tau}(\mathbf{X}^*)},
and var_between is the variance of the set of \eqn{\hat{\tau}(\mathbf{X}^*)}. For more details,
see Brantner et al., (2024).

The predict function can only be run on \code{cate} objects with variance estimates available. As
long as that criteria is met, any method used to estimate the CATEs across multiple studies can
then be used to create prediction intervals using this approach.

The \code{newdata_tbl} can consist of any set of profiles for which the researcher would like to
create prediction intervals. The assumption with this approach is that the profiles in the target
setting can be found within the covariate distribution of at least one of the studies used to
estimate the CATE (i.e., positivity). CATE prediction intervals will be constructed separately
for each covariate profile, represented by a row in the newdata_tbl.
}
\examples{
# inst/examples/example-predict_cate.R

new_dummy_tbl <- tibble::tribble(
  ~tx, ~var1,   ~var2,   ~var3,    ~var4,   ~var5,   ~response,
  0,  -0.728,   1.66,   -0.213,   -0.0621,  0.252,   1.26,
  0,  -0.464,   0.230,   0.579,   -1.17,    0.0513, -0.235,
  0,   1.52,   -0.792,   0.951,    0.869,  -0.312,   1.42,
  0,   0.410,  -0.945,  -1.49,    -1.87,   -1.35,   -3.10,
  1,  -0.223,   2.20,    0.0242,   0.419,   0.333,   3.56,
  1,  -0.187,  -0.439,  -0.557,   -0.0679,  0.497,   0.0518,
  1,   1.41,   -0.830,  -0.306,    1.12,    0.648,   2.61,
  1,  -0.437,   1.22,    1.18,    -1.63,   -1.09,    0.833,
)

cate_object <- estimate_cate(
  trial_tbl = dummy_tbl,
  estimation_method = "causalforest",
  aggregation_method = "studyindicator",
  study_col = "studyid",
  treatment_col = "tx",
  outcome_col = "response"
)

predict(cate_object, new_dummy_tbl)
}
\references{
Brantner CL, Nguyen TQ, Parikh H, Zhao C, Hong H, Stuart EA. Precision mental health: Predicting
heterogeneous treatment effects for depression through data integration. \emph{Arxiv}.
}
